Removes f32 and rename f64 and i32 to int and float

+++ src/Token.zig
@@ -76,8 +76,9 @@ pub const TypeValue = enum {
     u16,
     u32,
     u64,
-    f32,
-    f64,
+
+    int,
+    float,
     bool,
     void,
     str,
